Move FileSizeFormatter to module :libraries:androidutils
