+++ features/messages/impl/build.gradle.kts
@@ -34,7 +34,6 @@ dependencies {
     anvil(projects.anvilcodegen)
     api(projects.features.messages.api)
     implementation(projects.libraries.androidutils)
-    api(projects.libraries.androidtools.api)
     implementation(projects.libraries.core)
     implementation(projects.libraries.architecture)
     implementation(projects.libraries.matrix.api)
@@ -65,7 +64,6 @@ dependencies {
     testImplementation(libs.molecule.runtime)
     testImplementation(libs.test.truth)
     testImplementation(libs.test.turbine)
-    testImplementation(projects.libraries.androidtools.test)
     testImplementation(projects.libraries.matrix.test)
     testImplementation(projects.libraries.dateformatter.test)
     testImplementation(projects.features.networkmonitor.test)
