+++ features/preferences/impl/src/main/kotlin/io/element/android/features/preferences/impl/tasks/ComputeCacheSizeUseCase.kt
@@ -18,7 +18,7 @@ package io.element.android.features.preferences.impl.tasks
 
 import android.content.Context
 import com.squareup.anvil.annotations.ContributesBinding
-import io.element.android.libraries.androidtools.api.FileSizeFormatter
+import io.element.android.libraries.androidutils.filesize.FileSizeFormatter
 import io.element.android.libraries.androidutils.file.getSizeOfFiles
 import io.element.android.libraries.core.coroutine.CoroutineDispatchers
 import io.element.android.libraries.di.ApplicationContext
