+++ plugins/src/main/kotlin/extension/DependencyHandleScope.kt
@@ -77,7 +77,6 @@ private fun DependencyHandlerScope.addImplementationProjects(
 fun DependencyHandlerScope.allLibrariesImpl() {
     implementation(project(":libraries:androidutils"))
     implementation(project(":libraries:deeplink"))
-    implementation(project(":libraries:androidtools:impl"))
     implementation(project(":libraries:designsystem"))
     implementation(project(":libraries:matrix:impl"))
     implementation(project(":libraries:matrixui"))
